Fix issue with help being triggered inadvertedly
Help was supposed to be triggered only when no parameters are passed.
However using program.args.length to determine that is aparently wrong,
since that's empty when the options are provided before the command.

Using `process.argv.slice(2).length` is also shown in the commander.js
README and solves the issue.