Add script to publish beta package
Packages are tagged with the current git commit, since NPM does not allow
uploading a package with the same version twice.

Note that this script uses npm to publish, since yarn is a little more
difficult to automate. For example it doesn't allow to prevent the new
package.json to be committed.