Use Lerna to manage dependencies between Maji packages
This simplifies the project setup.
In the future we could utilise Yarn workspaces to prevent node_modules
getting installed in multiple packages, but the downside of this is that
you need to have a package.json with `private: true` in the root of the
repo so it changes the whole repo layout.