Don't clear app mount point
This was added back at a time when we were trying to
mimic React, which does this (while preact appends).
However back then we were not attaching directly to
the <body  />. Now that we are, this doesn't make
sense any more, as it removes from the DOM other tags
that are added to the body.